requirements.txt
aiofiles==25.1.0
fastapi==0.129.0
google-genai==1.63.0
httpx==0.28.1
orjson==3.11.7
pydantic-settings==2.13.0
python-dotenv==1.2.1
python-multipart==0.0.22
tenacity==9.1.4
uvicorn==0.41.0
websockets==15.0.1
sqlalchemy[asyncio]
asyncpg
pydantic

    